requirements.txt
#

cassandra-driver>=3.0
